Address feedback, make `InstanceBuilder` private
